requirements/development.txt
# Local development dependencies go here
-r base.txt

ipdb==0.10.3

django-debug-toolbar==1.9.1
django-extensions==1.9.9